// src/app/model/localidade.ts
export interface Localidade{
    cdLocalidade: number;
    dsLocalidade: String;
    dtRegistroNire: String;
    flSede: boolean;
    nmBairro: String;
    nmComplemento: String;
    nmEnderecoRodape: String;
    nmLogradouro: String;
    nmMunicipio: String;
    nmUf: String;
    nrCaixaPostal: String;
    nrInscEstadual: String;
    nrInscMunicipal: String;
    nrLogradouro: String;
    nrTelefone: String;
    situacaoLocalidade: boolean;
    vlCep: String;
    modificationTime: String;
    modifiedByUser: String;
    creationTime: String;
    createdByUser:String;
    CNPJ:String;
    nrNIRE:String;
}